ml-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.3
torch==2.11.0
torchvision==0.26.0
mediapipe==0.10.9
opencv-python==4.13.0.92
pillow==12.2.0